requirements.txt
Flask
Pillow
rembg
# Note: rembg may require torch (PyTorch) on some systems. If install fails, install PyTorch first:
# pip install torch torchvision
